Move user settings handling code to a module
