+++ rust/agama-lib/src/users/proxies.rs
@@ -1,0 +1,57 @@
+//! D-Bus interface proxies for: `org.opensuse.Agama.Users1.*`
+//!
+//! This code was generated by `zbus-xmlgen` `3.1.0` from DBus introspection data.`.
+use zbus::dbus_proxy;
+
+#[dbus_proxy(
+    interface = "org.opensuse.Agama.Users1",
+    default_service = "org.opensuse.Agama.Users1",
+    default_path = "/org/opensuse/Agama/Users1"
+)]
+trait Users1 {
+    /// RemoveFirstUser method
+    fn remove_first_user(&self) -> zbus::Result<u32>;
+
+    /// RemoveRootPassword method
+    fn remove_root_password(&self) -> zbus::Result<u32>;
+
+    /// SetFirstUser method
+    fn set_first_user(
+        &self,
+        full_name: &str,
+        user_name: &str,
+        password: &str,
+        auto_login: bool,
+        data: std::collections::HashMap<&str, zbus::zvariant::Value<'_>>,
+    ) -> zbus::Result<(bool, Vec<String>)>;
+
+    /// SetRootPassword method
+    fn set_root_password(&self, value: &str, encrypted: bool) -> zbus::Result<u32>;
+
+    /// SetRootSSHKey method
+    #[dbus_proxy(name = "SetRootSSHKey")]
+    fn set_root_sshkey(&self, value: &str) -> zbus::Result<u32>;
+
+    /// Write method
+    fn write(&self) -> zbus::Result<u32>;
+
+    /// FirstUser property
+    #[dbus_proxy(property)]
+    fn first_user(
+        &self,
+    ) -> zbus::Result<(
+        String,
+        String,
+        String,
+        bool,
+        std::collections::HashMap<String, zbus::zvariant::OwnedValue>,
+    )>;
+
+    /// RootPasswordSet property
+    #[dbus_proxy(property)]
+    fn root_password_set(&self) -> zbus::Result<bool>;
+
+    /// RootSSHKey property
+    #[dbus_proxy(property, name = "RootSSHKey")]
+    fn root_sshkey(&self) -> zbus::Result<String>;
+}
